Refactoring Location Services in Map and Settings
